requirements.txt
Django==5.1.1
djangorestframework==3.14.0
pandas==1.5.3
openpyxl==3.0.10  # For reading Excel files
